Add a WASM browser module with fetch() available
